# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Afrik'Fid API"
  version: "1.0.0"
  description: |
    Passerelle de paiement B2B multi-pays avec fidélisation intégrée pour l'Afrique francophone et anglophone.

    ## Authentification

    L'API supporte deux mécanismes d'authentification :

    **1. JWT Bearer** — Pour le tableau de bord admin et marchand
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtenir un token via `POST /auth/admin/login` ou `POST /auth/merchant/login`.

    **2. API Key** — Pour les intégrations marchand (paiements, confirmations, statuts)
    ```
    X-API-Key: af_pub_xxxxxxxxxxxx
    ```
    Ajouter `X-Sandbox: true` pour utiliser les clés sandbox et simuler les paiements.

    ## Modèle de fidélité X/Y/Z

    Chaque paiement est réparti entre trois parties :
    - **X%** — Commission marchand (remise accordée au client ou gardée)
    - **Y%** — Rebate client (selon statut loyalty : OPEN=0%, LIVE=5%, GOLD=8%, ROYAL=10%)
    - **Z%** — Commission plateforme Afrik'Fid

    ## Environnements

    | Environnement | Base URL | Clé |
    |---|---|---|
    | Production | `https://api.afrikfid.com/api/v1` | `af_pub_xxx` |
    | Sandbox | `https://api.afrikfid.com/api/v1` + `X-Sandbox: true` | `af_sandbox_pub_xxx` |
    | Local | `http://localhost:4001/api/v1` | Clé de test |

  contact:
    name: "Support Afrik'Fid"
    email: support@afrikfid.com
  license:
    name: Propriétaire
servers:
  - url: http://localhost:4001/api/v1
...
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/admin/login ou /auth/merchant/login
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API publique du marchand (af_pub_xxx en prod, af_sandbox_pub_xxx en sandbox)
    HmacSignature:
      type: apiKey
      in: header
      name: X-AfrikFid-Signature
      description: |
        Signature HMAC-SHA256 optionnelle pour l'intégrité des requêtes (CDC §4.5.2).
        Format : `sha256=<hex>`
        Calcul : `HMAC-SHA256(api_key_secret, raw_request_body)`
        En mode `REQUIRE_HMAC_SIGNATURE=true`, ce header devient obligatoire sur `/payments/initiate` et `/payments/{id}/refund`.

  # ─── Schemas ───────────────────────────────────────────────────────────────
  schemas:
...
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        reference: { type: string, example: "AFD-1710000000-ABC123" }
        merchantId: { type: string, format: uuid }
        clientId: { type: string, format: uuid, nullable: true }
        grossAmount: { type: number, example: 10000 }
        currency: { type: string, example: XOF }
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        paymentMethod:
          type: string
          enum: [MOBILE_MONEY, card]
        paymentOperator:
          type: string
          example: ORANGE
          nullable: true
        operatorRef: { type: string, nullable: true }
        clientLoyaltyStatus:
          type: string
          enum: [OPEN, LIVE, GOLD, ROYAL]
        merchantRebatePercent: { type: number }
        clientRebatePercent: { type: number }
        platformCommissionPercent: { type: number }
        merchantRebateAmount: { type: number }
        clientRebateAmount: { type: number }
        platformCommissionAmount: { type: number }
        merchantReceives: { type: number }
        rebateMode: { type: string, enum: [cashback, immediate] }
        initiatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time }
...
paths:
...
  /reports/daily:
    get:
      tags: [Reports]
      summary: Rapport journalier
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date, example: "2026-03-12" }
        - in: query
          name: merchant_id
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Résumé du jour
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string }
                  summary: { type: object }
                  byMethod: { type: array, items: { type: object } }
                  byLoyaltyStatus: { type: array, items: { type: object } }
  /reports/overview:
    get:
      tags: [Reports]
      summary: KPIs globaux (30 derniers jours par défaut)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: period
          schema: { type: string, default: "30d", example: "7d" }
          description: "Format: Nd (ex: 7d, 30d, 90d)"
      responses:
        '200':
          description: KPIs et métriques globales
          content:
            application/json:
              schema:
                type: object
                properties:
                  kpis: { type: object }
                  topMerchants: { type: array }
                  loyaltyDistribution: { type: array }
                  dailyVolume: { type: array }
                  merchantCount: { type: integer }
                  clientCount: { type: integer }
                  period: { type: string }
  /reports/transactions:
    get:
      tags: [Reports]
      summary: Export des transactions (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: merchant_id
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { type: string, enum: [pending, completed, failed, refunded] }
        - in: query
          name: loyalty_status
          schema: { type: string, enum: [OPEN, LIVE, GOLD, ROYAL] }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items: { $ref: '#/components/schemas/Transaction' }
                  total: { type: integer }
  /reports/exchange-rates:
    get:
      tags: [Reports]
      summary: Taux de change (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  rates:
                    type: array
                    items:
                      type: object
                      properties:
                        fromCurrency: { type: string }
                        toCurrency: { type: string }
                        rate: { type: number }
                        source: { type: string }
                        updatedAt: { type: string }

    put:
      tags: [Reports]
      summary: Mettre à jour un taux de change (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_currency, to_currency, rate]
              properties:
                from_currency: { type: string, example: XOF }
                to_currency: { type: string, example: EUR }
                rate: { type: number, example: 0.00152 }
      responses:
        '200':
          description: Taux mis à jour
  /reports/overview-normalized:
    get:
      tags: [Reports]
      summary: KPIs normalisés en EUR
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: period
          schema: { type: string, default: "30d" }
      responses:
        '200':
          description: Volumes normalisés en EUR par devise
          content:
            application/json:
              schema:
                type: object
                properties:
                  bycurrency:
                    type: array
                    items:
                      type: object
                      properties:
                        currency: { type: string }
                        volume: { type: number }
                        tx_count: { type: integer }
                        volumeEUR: { type: number }
                  totalVolumeEUR: { type: number }

  # ── FRAUD ─────────────────────────────────────────────────────────────────
...
  /reports/transactions/pdf:
    get:
      tags: [Reports]
      summary: Export PDF des transactions (admin — CDC §4.6.1)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: merchant_id
          schema: { type: string, format: uuid }
        - in: query
          name: status
          schema: { type: string, enum: [success, failed, pending, expired] }
      responses:
        '200':
          description: Fichier PDF des transactions
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '501':
          description: PDFKit non installé
  /reports/transactions/excel:
    get:
      tags: [Reports]
      summary: Export Excel des transactions (admin — CDC §4.6.1)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: merchant_id
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Fichier Excel (.xlsx) des transactions
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '501':
          description: ExcelJS non installé
  /reports/by-country:
    get:
      tags: [Reports]
      summary: Rapport de revenus par pays/zone (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Volume et commissions par pays et zone monétaire
          content:
            application/json:
              schema:
                type: object
                properties:
                  byCountry:
                    type: array
                    items:
                      type: object
                      properties:
                        country_id: { type: string }
                        country_name: { type: string }
                        zone: { type: string, enum: [UEMOA, CEMAC, EAC] }
                        currency: { type: string }
                        transaction_count: { type: integer }
                        total_volume: { type: number }
                        total_commission: { type: number }
  /reports/refunds:
    get:
      tags: [Reports]
      summary: Liste des remboursements (admin)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: status
          schema: { type: string, enum: [pending, completed, overdue] }
      responses:
        '200':
          description: Liste paginée des remboursements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        transaction_id: { type: string, format: uuid }
                        amount: { type: number }
                        currency: { type: string }
                        status: { type: string }
                        reason: { type: string }
                        created_at: { type: string, format: date-time }
                  total: { type: integer }

  # ─── Clients — transactions & wallet ────────────────────────────────────────
...